Server/requirements.txt
tensorflow==2.0.0
keras==2.3.1
h5py==2.10.0
numpy==1.18.1
opencv-python-headless==4.2.0.32
pillow
fastapi==0.61.1
gunicorn
uvicorn==0.11.8
python-multipart
requests==2.24.0